Change gtk.Window to gtk.Dialog in gpodder.glade (bug 95)
This is a new changeset based on the patches available in
Bugzilla bug 95, which convert all gtk.Window widgets in
the gladefile (except for the main window) to gtk.Dialog
widgets, so the Esc key can be used to close dialogs, etc.

Thanks to Jérôme Chabod who did some experiments on this.