Detect and report HTTP errors when downloading (bug 106)
The FancyURLopener by default ignores HTTP errors and
downloads the HTTP error page without letting the calling
code know. We override the method and report the error
in our subclass, so we can provide feedback to the user
in the logfile and in a message dialog.

This bug has been reported by Markus Golser