Improve slash and duplicate episode name handling in gdfs (bug 170)
Replace the forward slash ("/") in the channel and episode
title with a dash ("-") to prevent episode title text from
being stripped.

Apply a modified version of a patch by Mel Jay to fix problems
with duplicate episode names, and try to find free names when
the filename already exists.

Thanks to Mel Jay for reporting this bug in gPodder Bugzilla