Performance fix for the main episode list.
When the list of episodes for a channel was displayed, all episodes
were loaded with a single query, then loaded separately by one
inside podcastItem.iter_set_downloading_columns().  This patch
fixes this problem, the TreeModel is built with only one SQL query.

On the same machine a TreeModel for 50 episodes was displayed in
1.915531 seconds, with this patch: 0.236303 (8 times faster).