Stop looking for episodes marked as deleted that still exist in fs.
If an episode is marked as deleted but the corresponding file somehow
still exists in the file system, this code was "resurrecting" the
episode, and vice-versa for downloaded episodes that lack their files.
Questionable behaviour, because a situation like that can only mean
an error somewhere, and this code was just making it harder to track
the real problem, and added unnecessary FS interaction.