YouTube integration.
Links to YouTube profiles are converted to the corresponding
RSS feeds (http://www.youtube.com/rssls), which aren't available
with the standard feed discovery.

Normal links to YouTube enclosures (*.swf) are now on-the-fly
replaced with links to high quality MP4 videos.  Apparently
links to real enclosures are not permanent, so they can't be
saved in the database.