Improved startup time by minimizing the number of SQL queries.
channels_to_model() was issuing 3 SQL queries per channel to
get episode statistics (the number of downloaded, new and unplayed
episodes).  This is now done with one query for all channels.
3 temporary views are created when the database is first open.

This not only affects the startup time, but also the time required
to finish a download, change settings -- everything that renews
the list of channels.