Fix size calculation with sync_only_not_played turned on
GPodder don't respect option 'Sync only not played episodes'
so when you click sync button it says something like 'You
need 1.5 GB of free space on your device' because it counts
size of all episodes, not only new ones.

This patch fixes this problem by counting only episodes that
are going to be copied to the MP3 player.