Remove URL entry box above the podcast list
Remove the URL entry box and make the "Subscribe" menu
item and the "Ctrl+L" keystroke always pop up a dialog.

Add an advanced option for users that want to have
the URL entry box back (for now). We might decide
that we will completely remove this feature in the future.