Expose system default video and audio player in UI
In addition to selecting audio and video players, we
now support "default" as a way of having a good default
that we can pass to a utility function that can later
be made cross-platform more easily.

Also cleaned up some UI issues with the video player,
so the label of the custom command is shown and the
combo box has the correct size.