Show information dialog while downloading feeds
Make sure that the user gets notified while we are
downloading feed data, so that the UI does not lock
up or look broken/slow when adding many feeds (e.g.
on initial OPML import).

Also add a GTK Mainloop iteration when blocking the
main thread for adding podcasts to not make the UI
freeze, but update it in every loop iteration.