New config and download folder locations on Maemo (bug 191)
Config files are moved from /media/mmc2/gpodder/ to
~/.config/gpodder/ when migrating from old versions,
so the N8x0 backup tool backups our config, too.

As a result, we move the download directory from
/media/mmc2/gpodder/downloads/ to /media/mmc2/gpodder/
and add detection for moving this to mmc1, the user's
homedir or an externally attached USB mass storage device.

Thanks to Dan Ramos for the bug report.