Crop tooltips and only show on first column (bugs 249, 186)
Crop episode list tooltips to a maximum of 400 characters.

Only show tooltips in main window when hovering over the
first column (podcast cover in the podcast list, and the
status icon in the episode list).

This should make the tooltips less annoying.

Thanks to Andrew Bennett and Steve McCarthy for reporting.