Rename "close" to "cancel" in Add new podcast dialog (bug 265)
Problem description and justification:

When adding a podcast, there are two buttons -
close and add. I suggest from a useability point
of view, that you rename 'close' to cancel which
people are more used to seeing – I kept hitting
this when I actually wanted to do was 'close the
dialogue and continue – ie add).  I then had to
go back and reenter the subscription information again.

Thanks to Vic for reporting this bug on Bugzilla.