Do not forget custom command when opening preferences (bug 289)
Due to the change of the list of items in the players
list, we incorrectly assumed the last item is the custom
command item.

Now we do some more sophicistcated searching for the custom
command item which also fixes this bug.

Thanks to S. Rust for reporting this bug on bugs.gpodder.org.
(cherry picked from commit 3e342347a830a2893247070e97713d152c01ec15)