Add a message to AttributeError in config.py
If you happen to try to use a non-existing config
variable inside a try-catch where you catch
Exception and then log its message, the resulting
error message isn't going to be very helpful when
the caught Exception itself has no message.

This adds a simple message to the AttributeError
which is raised by config.py if a non-existing
config variable is attempted to be used.