Only show sync window when needed (bug 294)
The synchronization window should only show
up when episodes are really deleted from the
device, not while the selection or confirmation
is still going on.

This patch fixes this issue and also makes the
deletion from the device threaded, so the UI
should be more responsive during delete actions.

Thanks to Silvio Sisto for reporting this bug.