Add option to skip disk sync after transfer (bug 338)
Advanced configuration option to not call the "sync"
system utility to flush caches to disk.

We usually do this, so that basic users may unplug
their devices right after gPodder shows "sync done".

For advanced users who know to correctly unmount their
devices, this is annoying, so let them disable it.

Thanks to Christian Mertes for requesting this feature.