Avoid a recursion situation in file_exists()
There was a possible infinite recursion situation
in the check for file existence. This patch fixes
this by allowing the local_filename() function to
accept a "check_only" parameter that will not try
to rename/fix wrong filenames, but only check for
their existence.

Thanks to Marco Villegas for reporting this bug.

(PS: Writing log messages that magically auto-fit
a block paragraph is fun :)