Add "gpodder-backup" command-line utility
This can be used for creating and extracting tarballs
that contain the complete gPodder configuration data
and downloads.

It is also useful for automatically packing up all the
necessary files that developers would need to reproduce
bug reports, and leaving out the (potentially big) data
of downloaded files.

Manual page included. Use it! :)