Avoid problems with wrong timestamps (bug 352)
I've search a little and it seems that it comes
from podcasts that are on my iPod and that don't
have a released date.

I've just attached my version of util.py. I've
added a try/except, so if the timestamp is not None,
but is incorrect, the exception is catched and None
is returned. It works for me.