Fix channel foldername that looks like an URL
This bug has been introduced by fixing another bug,
and so we're now checking if the folder name starts
with the "sanitized URL"-type folder name. If it
does, we assume that the folder name was created
when the podcast title was not yet available, and so
we're obliged to try to find a better name now.