Only update filename when extension changes (bug 363)
Only update the filename of an episode as a result of
changed mimetypes when the changed mimetype also changes
the file extension.

Also fix a small bug that omitted the extension when
searching for a "proper" filename and a bug that happens
when the filename extension is missing from a file.

Thanks to Ville-Pekka Vainio for reporting this bug.