Add support for unit testing and doctesting
New module "gpodder.unittests" that will run all
unit tests and doctests (currently only gpodder.util)
for gPodder. The tests can easily be run with a

    make unittest

in the source tree or with the following command
when gPodder is installed onto the system:

    python -m gpodder.unittests

Also added a (currently unused) helper function to
gpodder.util that already has unittests prepared.