Detect web-based authentication pages for Wifi (bug 365)
When a user connects to a public Wifi hotspot that has
some kind of HTTP-based authentication scheme (or even
just a welcome page of some kind), do not update the
feed but detect this and ignore the update.

Later improvements can show a notification and open
the default webbrowser.

Thanks to Christian Mertes for reporting this bug.