More doctests; test "find_mount_point" (using MiniMock)
Add doctests for the find_mount_point() utility method
and make it more robust while finding shortcomings of
the current implementation by unexpected behaviour that
has been checked with the doctest and mock objects.

Note the unittest requirements (minimock + coverage) in
the README file and rm the ".coverage" file on "make clean".