Doctest for util.username_password_from_url
Add doctests for the username/password from URL
function and raise a ValueError when the input
data is invalid (wrong type or format - see below).

Had a look at RFC1738 (Section 3.1) how username
and password are specified in URLs. According to
that RFC, "@", ":" and "/" have to be encoded in
the username and password field.