Sync episodes ordered by publishing date (bug 388)
The episodes that are queued to sync are now ordered
by the pubDate attribute, so that MP3 Players (like
the Sansa Fuze) sort podcasts in the correct order.

Some podcasts such as NPR's Science Friday podcast do
break up a single episode in several parts, which has
led to parts playing in reverse order.

Thanks to Abraham D Smith for the bug report and an
initial patch to solve this issue.