Fix channel deletion in CLI mode (-d)
Channel deletion did not work in CLI mode, because
we were only saving the OPML file, and did not
remove the channel from the database. This patch
fixes it by calling delete() on the channel object.

Thanks to Markus Golser (elmargol) for testing
the CLI mode and bringing this to my attention.