Fix translation issues with gtk.Builder
This should make the gtk.Builder translations work
on Linux and other free operating systems with
'bindtextdomain' in the locale module. For other
systems (it seems like Windows does not have that
function in the locale module), a small C module
is provided that does the same thing. It will be
used if available.