Fixed version of redirect_internal in DownloadURLOpener (bug 465)
Using wireshark you can see that gPodder gets stuck after it receives
the first 302 from the server. Wget on the other hand handles it
perfectly (there are 2 to 3, 302 responses in total).

Adding a Range header to the request will make the download fail,
this patch fixes this bug by providing our own version of
redirect_internal.