Only offer new episodes of updated feeds for download (bug 473)
When a single feed was updated, the offer to download new
episodes was based on *all* feeds, not just the updated one.
This is now fixed, and the search for new episodes is now
restricted to only the channel(s) that has/have been updated.

Thanks to scruffythinking for reporting this bug.