Add free disk space detection on Windows (bug 468)
I'm attaching a patch against the current git tree which should fix
this, based on the information here:

http://mail.python.org/pipermail/python-list/2003-May/203223.html

Note that the user will need to have the Python win32 extensions
installed - see http://sourceforge.net/projects/pywin32/ - and I have no
idea how to include these in the frozen Python binary which is included
in the win32 distribution of gpodder.  The patch reverts back to the old
behavior if it can't find the win32api.