Only warn user when free space is insufficient (bug 537)
When syncing episodes to a player, do not abort the process
of syncing to devices that might not report the free space
correctly (like obexfs), but ask the user to confirm the
(probably risky) operation of trying to fill the device with
data.

Thanks for phq for the suggestion in bug 537.