Split out GTK+-related parts of gpodder.config module
Create a new UIConfig class that subclasses the normal
Config class that takes care of the basic configuration
work.

The UIConfig class adds support for registering GTK+ UI
elements with the configuration object.

Additionally, the ListStore (TreeModel) code is split
into a new class ConfigModel that takes care of formatting
the configuration data for display in GTK+ applications.