Remove "E-Mail subscription list" menu item (bug 563)
There have been problems with this feature, and it
is only supported on the Linux Desktop right now,
anyway, so the feature will have to go.

The proposed replacement is a manual file-based export
of the subscription list as OPML and then manually
sending this file via E-Mail.

Thanks to hughitt1 for reporting this "bug".