Fix cover file name in backup and sync (bug 591)
When adding new feeds, the new filename ".cover"
is not detected correctly when syncing files to
an iPod or backing up the download folder.

This patch fixes it by first trying the old name
("cover") and then the new one (".cover") if the
old one does not exist.

Thanks to Romain Janvier for reporting this bug.