Add "fb:" prefix shortcut for feed URLs
It's now possible to have some prefix shortcuts
when adding URLs. This should make the life of
lazy typists easier, especially on mobile devices.

At the moment, the only available prefix is "fb:".
If you want to add a FeedBurner URL, simply take
the name of the podcast and prefix it with "fb:".

Added some doctests to normalize_feed_url.