Fix alignment in Maemo 4 preferences (Maemo bug 6205)
Depending on the string length (eg. in translations),
the preferences dialog's picker buttons had their
text misaligned, which looked ugly. This patch fixes
the issue by making the labels fill the whole width
and adding a padding to the left of the label.

Thanks to Tomasz Dominikowski for reporting this bug.