Do not offer auto-upload to mygpo (bug 671)
Do not ask the user to auto-upload the subscription
list to my.gpodder.org on quit. We are going to
rework most of the client code anyway, at which
point we will have a setting in the preferences.

As a consequence of this patch, Maemo bug 5268 has
also been fixed (as it was caused by this dialog).