Folder names must not end in dots (bug 600)
On Windows, a folder name like "Abc..." is
saved as "Abc", and trying to create such a
folder name works, but the trailing dots are
removed. When trying to download files to such
a folder (while gPodder still believes the
folder has dots at the end), this won't work.

To fix this bug, don't allow trailing dots in
the folder name of podcasts by stripping them.

Thanks to Yves for reporting this bug.