Spawn new threads for download limit (bug 693)
When the "download limit" value is changed,
automatically start new threads if it makes
sense (when there are still tasks waiting
to be added and the amount of current threads
is less than the download limit).

Also let each thread check if it may continue
to accept tasks instead of using "cancelling".
This way, the number of threads more often is
in line with the download limit setting.

Thanks to "ad" for reporting this issue.