Remove filename discover code (bug 821)
The filename will be updated by the download
module whenever the download has finished and
the file_disposition header contains a proper
name. This is easier (and also gets rid of the
blocking operation during filename discovery)
than what we did before, which (I believe) also
led to an error that Shane Kerr was seeing in
gPodder bug 821.