Don't mark old episodes as new on update (bug 340)
Another attempt at finally fixing bug 340, but this time I'm pretty sure
that it works, because we are more selective now.

Whenever a new episode appears in the feed, its date is compared to the
date of the newest-known episode in the same feed. If it's more than one
week older than the newest-known episode, it's marked as old by default.

The "one week" value is just a randomly-chosen one to allow for border
cases (like a daily podcast re-issuing an episode of the last few days)
to work correctly and not cause "missing episodes".