Subscription handling for Mac OS X (bug 959)
Register for Apple Events of type 'open document' so that one
can drag and drop an OPML file to the gPodder icon.
Register for Apple Events of type 'GURL' so that one can
choose gPodder when subscribing to a podcast in Firefox.

Drag and drop events work only at startup, I don't know why.