Refactor database module and add new columns
Clean-up the database module for easier adding and
removal of columns to the podcast and episode tables.

Add time-related information (total time, position and
timestamp for the last position update) to episodes.

Remove the "deleted" column for podcasts and make sure
to remove all "deleted" podcasts and their episodes when
upgrading from an older version of gPodder.