Fix credentials-with-space bug (bug 1065)
Spaces in username or password for HTTP auth were
quoted in a wrong way (quote_plus), which made the
rest of the code not decode the URL correctly.

This patch makes quoting of spaces work correctly
and produce URLs that have correct semantics. Unit
tests were also added to verify this special case.