MTP: Configurable podcast folders (bug 1098)
I've recently switched from an iPod to a Sony Ericsson mobile phone,
which is accessible as an MTP device. It has support for podcasts, but
it only recognizes tracks as podcasts if they're in certain folders,
namely:

    Music/podcast
    Video/podcast
    Picture/podcast

Tracks from these folders show up in the "Podcasts" menu, are
highlighted until played and per-file playback positions are remembered.

Attached is a patch which adds configurable folders for audio, video and
image podcasts. There is also a new option to create folders per podcast
(e.g., Music/podcast/No Agenda). The complete list of added config
options is:

    mtp_audio_folder (str)
    mtp_video_folder (str)
    mtp_image_folder (str)
    mtp_podcast_folders (bool)