Fix a podcast list update bug with "All episodes"
If the view filter on the episode list is set to
"Downloads only", marking an episode as not new
in the "All episodes" view will not update the
podcast in the podcast list due to the selection
being removed when the episode list is updated.

The fix is to update the podcast list before the
episode list, leaving the selection intact until
the podcast list has been updated.