Fix inconsistent tray icon behaviour (bug 1125)
In window managers like Metacity or KWin4, the
main window wasn't correctly moved to the current
virtual desktop when the tray icon was clicked.

To work around this, the window is now hidden and
then re-shown when the tray icon is clicked, which
fixes the issue with the affected window managers.