Remove *_played_dbus from config
This avoids problem with "played" status tracking
on Maemo. With this change, we always mark an
episode as "played" as soon as it is loaded into
a media player, but due to the MAFW time tracking
we don on Maemo 5, we can now determine the "finished"
state through other means, anyway.