Config: Hide positional values in UI editor
There is no need to see/edit the window position and
paned position values in the graphical config editor.

The positional (window status / paned position) config
values get a "_" prepended to their name, and all values
starting with "_" are hidden in the UI editor. They can
still be edited in the "Settings" file, but there should
really be no need to edit them.