Rewrite check for received actions (bug 1391)
When processing notifications from a media player about 'play' actions,
we check to see that the previous update didn't occur at a time later
than 'now', and only if it didn't do we update with the new action. This
seems unnecessary (and is therefore confusing): the only way that a
previous action could have happened before 'now' is if one or the other
of the timestamps being compared is incorrect --- in which case it's
probably still correct to update with the new action?

If the check was put in for a specific reason, I think it should be
clarified with a comment; otherwise, I think it can be safely removed.
The attached patch asserts that the condition is true (I'll run with
this for a while just to see what happens), though I would just as soon
get rid of that, too.