Ignore image enclosures for audio/video in Media RSS (bug 1430)
We have only checked if audio and/or video is available in
normal enclosures. This check is now extended to Media RSS
items, because there are feeds that have the image as
enclosure and the media file as Media RSS content.

Example feed: http://www.rtl.fr/emission/a-la-bonne-heure.rss

(This commit has been merged from the "tres" branch)