Gtk UI: Toggle "new" status by clicking the icon (bug 1432)
Toggling the new status of episodes works when the clicked episode
is selected and the icon is clicked. Setting the new status also
works when multiple episodes are selected - the clicked episode is
toggled, and depending on the new status of the clicked episode,
the rest of the selected episodes is also updated with this status.