Avoid early deletion/collection of gPodderWelcome (bug 1021)
The gPodderWelcome Python object was deleted and/or collected
after it fell out of scope, while the Gtk dialog was still
displayed. Work around this by running a separate main loop
for the dialog, so that the Python object stays alive longer.

This bug has been haunting us for quite some time now, and it
seems like the object hasn't always been collected, making it
harder to track down the bug. But rejoice, we finally fixed it!