Custom feeds: Delete episodes when removed from feed
Up to now, the episodes from non-feed sources (Soundcloud
and XSPF) would pile up forever, because they were never
deleted from the database, even if they were removed from
the original feed. This is now fixed using the same mechanism
as the one for normal feeds: If an episode GUID is not found
in the feed, it is removed from the database.