QML UI: Scrubbing and other UI improvements
You can now scrub the playback buttons in the media
player to quickly seek around audio podcasts.

Also the graphics have been improved and a visual
preview of the playback bar is now implemented.

The color of the playback bar changes to blue to fit
the representation of playback (blue) in the UI.