Prefix elimination: Podcast title ' - ', ': '
This improves the prefix elimination code by also
checking if the episode title begins with the title
of the podcast, followed by either ' - ' or ': '.

While this is usually covered by the common prefix
elimination anyway, it's not covered by situations
where there are a few episodes that have a different
name (e.g. happens for the 43 Folders podcast now).