CLI: Improved output for "update" and "pending"
The update command now shows the sum of new episodes
in all updated episodes (this includes "new" episodes
that have been found in a previous update but not yet
marked as old or downloaded).

The pending command now shares the pending message
with the update command, and also uses color output
to highlight the podcast and episode names list.