gPodder 2.20.1 "Rosey the Robot" released
This is the first bugfix release for the 2.20.x series of releases,
containing fixes from the 3.x release series, with some customizations
to work well on Maemo 4 and Maemo 5.

Users who are not dependent on the support of Maemo 4, Maemo 5 or MP3
Player Sync in 2.20.x are recommended to upgrade to gpodder 3.x and not
stay on 2.20.x.