Restructure the source tree, modular installs
Restructure the source tree to be more like the
target installation directory (share/...).

Rewrite setup.py to only install a subset of the
available UIs and to be more flexible and cleaner
in general.

I hope this doesn't cause too much breakage, but
should result in a cleaner build/install system and
more happy developers and packagers in the long run.