Windows: Check for software updates (bug 1569)
The check can be carried out on Linux and OS X
as well, but it's mainly targetted at Windows
users, as Linux users will usually get updates
via package repositories. Only for the Gtk UI.

By default, the check is only enabled on Windows
and can be disabled manually via the config editor.