Podcast dialog: Section changing and Flattr bug (bug 1636)
The Flattr integration was causing the dialog to re-appear if
it is closed to quickly, because of its asynchronous nature,
so it has been ripped out for now.

Also, the section changing wasn't working, so instead of having
a comboboxentry, we now have a combo box and a button that will
open a text entry dialog and allow a new section name to be
entered. This will arguably also improve the user experience.