Basic Offline mode support (bug 107)
Detect network connections on Linux (still need to find a way to
implement it for Windows and OS X - on these platforms, this
implementation acts as if the connection is always available).

We might need to use this in more places in the code, but for starters
this should work nicely.

The Linux implementation assumes the "ip" command is available.