Linux: Detect PPP as Internet connection (bug 1702)
PPP connections on Linux have state "UNKNOWN", and are
therefore not detected as such. We fix this now by assuming
that all "UNKNOWN" connections are actually working Internet
connections, and filter the loopback interface (which is also
an "UNKNOWN" connection).

Thanks to Sudaraka Wijesinghe for the bug report, initial
patch and testing/feedback.