Fixing TEDtalks subtitle downloader
the introDuration parameter on tedtalks page has changed a little, now the
parameter appears using htmlentities, so I had to change the split for it to be
able to find the subtitles.
Also now the subtitles appears in seconds.
And I've provided 15 as a default value, since it's a good guess for modern
tedtalks (older ones have longer introductions)