Make online detection a bit smarter (bug 1730)
If we don't have "ifconfig" and "ip" assume we're online
(because we can't determine that we're offline), and if
detecting the network connection fails for some reason,
also assume that we're online. This avoid bogus "you're
offline" messages while still giving good feedback to
users who try to check for new episodes while offline.