Add option to flattr an extensions and to show documentation
If an extension author adds an flattr url to the extensions metadata
gPodder shows the "Flattr this" menu entry in the popup menu for an
extension in the preference dialog.
Alternative the extension author can add an url to the extensions
metadata to link to the authors donation page (Support the author).

Also the extension author can add an url to the extensions metadata
to link to the extensions user documentation.