QML UI: Highlight podcasts on Harmattan, but not on Sailfish
From user feedback, the colored text highlight in the podcast list
for podcasts with new episodes has been received better than the
new "triangle" indicator. So we'll now show the colow highlights on
Harmattan (with black background), whereas on Sailfish we still
don't show it (at least with the default background in the emulator,
the white text looks better than the colored text).