Gtk UI: Fix crash on OS X when truncating tooltip (bug 1825)
When displaying a long episode description in a tooltip in the
"New episodes available" dialog, gPodder cuts it at a certain
number of characters. However, Russian descriptions are regular,
non-unicode python strings, and gPodder may cut only a part of a
multi-byte UTF-8 sequence. Displaying such a string crashes
gPodder on OS X.

This patch fixes that by converting the description to a unicode
string if it's not unicode.