Gtk status icon: load icon by name except when actually drawing on it
The icon image quality degradation described in the previous changeset is
actually a result of loading by pixbuf, regardless of whether we then draw on
it or not. Interestingly, if the initial load is done by name, then subsequent
loads by pixbuf maintain the higher quality. In any case, in order to minimize
the quality impact, only load the image by pixbuf when actaully doing the
drawing. And the entire thing is still optional -- if download_progress_bar is
not set, the icon is only ever loaded by name.

--HG--
extra : rebase_source : 21233543a2e668bacdd43af1055c99d3e4b8e8e8