mpris listener: rename 'same_second' -> 'subsecond_difference'
'same_second' is a bit of a misleading name: it could be understood to check
whether or not two times are "in the same second" rather than whether two times
are "within a second of each other" (so, 1.999 and 2.001 might be thought to
fail this check).

Indeed, the function was originally implemented as "in the same second"; but
that implementation was incorrect in the context in which we use it, and the
implementation was fixed in [1]; unfortunately, the name was not changed as
part of that fix... Fixing the name, too, now.

[1] https://bitbucket.org/dov/gpodder-mpris-listener/commits/da2c5444fe83f6f7106714a6750a96d49d97a6a8

--HG--
extra : amend_source : 0da64760fe6926460641f9ef158bcd23963bf963