QML UI: Don't apply margin fix in landscape (bug 1915)
For some reason, we needed to apply a offset of 36 pixels, so that the
overlay sheet aligns with the window. However, this offset gives
problems in landscape (it also gives a visual glitch when the sheet is
closed, as it could be partially visible at the bottom of the window).
This fixes it by not applying the offset fix when the window is in
landscape mode or when the sheet is closed.

Thanks to Andi for reporting this issue.