Show covers directly after adding feeds (bug 1987)
When adding a new podcast feed a cover will get downloaded in the
background. Once it is finished gPodder.cover_download_finished() calls
PodcastListModel.add_cover_by_channel(), however by then a placeholder
cover will already be in the cache and the new cover will not be shown.

cover_download_finished() is the only caller of add_cover_by_channel(),
so generally clearing the cache when adding a new cover should be fine.