Podcast details dialog: Avoid re-showing on close
When the podcast details dialog is closed, it requests the
re-load of cover art. This re-load action might in some cases
cause the dialog to re-appear. To avoid this, only show the
dialog the first time the cover art is loaded, and additionally
disconnect from cover art signals before requesting on close.