YouTube Show Image Sometimes Low-Res
The link being parsed for the YouTube cover art image is sometimes only 100x100px, it appears to be inconsistent from show to show, sometimes 100x100, and sometimes 900x900px.  There is a link in the channel page source that appears to always have a 900x900px image.

This change looks for the "new" link first, to grab the higher quality image.  If not found it falls back to the original link.

Here are two examples of YouTube channels showing the links to the two different sized images:

The Ben Heck Show
https://www.youtube.com/user/thebenheckshow

Element 14
https://www.youtube.com/user/element14