Eliminate progress tab
As discussed in #878, this eliminates the top-level GtkNotebook and
moves the "Progress" tab into the right pane. Switching between the
episodes pane and the progress pane is achieved via a dedicated button
at the bottom of the left pane.

Eliminating the notebook gives slightly more vertical space to the
content which is beneficial when using gpodder on mobile devices. In
addition, without the wrapping notebook it'll get easier to implement
an adaptive UI that dynamically switches between the left, right and
both panes in the future.